tests/pyproject.toml
[project]
name = "org-dev-tests"
version = "0.1.0"
requires-python = ">= 3.11"
dependencies = [
    "pytest >= 8.3.0",
    "pytest-playwright >= 0.5.0",
    "requests >= 2.31.0",
]

[tool.pytest.ini_options]
testpaths = ["."]
pythonpath = ["."]
